// src/types/email.ts
export interface EmailRow {
  id: string;
  clientes: string;
  preco: string;
  dataBase: string;
  reajustado: string;
  fornecedor: string;
  medidor: string;
  consumo: string;
  perdas3: string;
  requisito: string;
  net: string;
  medicao: string;
  proinfa: string;
  contrato: string;
  minimo: string;
  maximo: string;
  faturar: string;
  cp: string;
  email: string;
  envioOk: string;
  disparo: string;
  dataVencimentoBoleto: string;
}